CONTACT-APP
Hola!!
This is a Contact Application using Java and Android Studio. It includes features such as  Adding a Contact, Deleting a Contact, Fetching Call History, Showing all Contacts, Adding to Favourite Contact.
The application starts with a Splash Screen and contains three fragments - History, Contacts and Favourite. We have used Recycler View to display contacts in the fragments.